requirements.txt
anthropic
langmem
langsmith
python-dotenv